requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
blosc=1.21.0=h8c45485_0
bottleneck=1.3.2=py37heb32a55_1
brotli=1.0.9=he6710b0_2
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py37h06a4308_0
cfitsio=3.470=hf0d0db6_6
charls=2.2.0=h2531618_0
cloudpickle=1.6.0=py_0
cudatoolkit=9.2=0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.7.2=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd3eb1b0_0
giflib=5.1.4=h14c3975_1
glib=2.69.0=h5202010_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
imagecodecs=2021.6.8=py37h581e88b_0
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py37h2531618_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=2.2.1=h2531618_0
libaec=1.0.4=he6710b0_1
libcurl=7.71.1=h303737a_2
libdeflate=1.7=h27cfd23_5
libedit=3.1.20210216=h27cfd23_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libwebp=1.0.1=h8e7db2f_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
libzopfli=1.0.3=he6710b0_0
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h295c915_1
matplotlib=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.2=he6710b0_1
networkx=2.6.2=pyhd3eb1b0_0
ninja=1.10.2=hff7bd54_1
numexpr=2.7.3=py37h22e1b3c_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
olefile=0.46=py_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.1=py37h8c16a72_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=8.3.1=py37h2c7a002_0
pip=21.2.2=py37h06a4308_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.6.0=py3.7_cuda9.2.148_cudnn7.6.3_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=0.24.2=py37ha9443f7_0
scipy=1.6.2=py37had2a1c9_1
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=he6710b0_0
sqlite=3.36.0=hc218d9a_0
threadpoolctl=2.2.0=pyhb85f177_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchvision=0.7.0=py37_cu92
tornado=6.1=py37h27cfd23_0
tqdm=4.62.0=pyhd3eb1b0_1
wheel=0.36.2=pyhd3eb1b0_0
xlrd=2.0.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zfp=0.5.5=h2531618_6
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
